Adds login/logout functionalities
Really proud of this one because I couldn't get it to work at... well,
work. The thing is the error message 'Please log in.' is added by the
filter, to the session ('cause the filter doesn't see the FacesContext)
and then, it's caught and added to the FacesMessages by the
LoginController. And the neat trick here is that, normally, the
constructor is called _after_ the p:messages tag is rendered (I'd love to
know the reason behind this one...). But if the message is caught in a
separate function, called by a f:viewAction tag (in a f:metadata tag), it
works as expected! So cool! (^_^)