Adds button to add Category in line page.
Tried to make Categories in-place editable but adding a category makes all
the in-place revert back to output state. So they're not.
TODO: add a way to delete a Categorisation and value verification before
saving (no empty categories, not twice the same category, amount must be
all positive or all negative, amount must match line amount...). And
TESTS!