WIP - Saved before major refactor
Ok, so the DAO layer (using Hibernate) has major problems: it uses 2
tables to save Cateogorisations and when a Category is modified, the
reference is lost in all Lines. The last one is too big to ignore. So
I'm going to refactor it to use direct SQL. But before I do that, I must
have tests for everything else, to be sure that nothing gets broken in
the process. No regression.